pyproject.toml
[project]
name = "gordon-in-your-kitchen"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-openai>=0.3.35",
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
]
